KL-162/refactor: separate initialize state
- state의 사전 설정을 useInitializeState로 분리하였습니다.
